Initial import.
This contains an almost complete implementation of the builder part of
the BSON spec, with a decent API and proper error handling. Also
included is the start of the test cases.

BSON parsing is not implemented in any shape or form yet, though.

Nor are there any tools supplied to build the sources.