Slightly improved tests.
The idea - in the long run - is to build various bson objects, and
dump them to stdout.

The run-tests.sh script will then compare the output to a known-good
set. A python script to generate the known-good set will be included
(but the generated file will appear in EXTRA_DIST, so that non-git
checkouts will not need python).