Even more improved test suite.
Added tests/gen-known-good.py, a python script that will generate
known-good output (using the bson module from pymongo >= 1.9).

The run-tests.sh script was modified to use the output of the above
script to test against. And the test program was modified to put the
boolean values into the BSON object in the same order Python does.