tests: Transition to python-based testing
Instead of generating a known-good output and comparing against that
(which is vulnerable, due to the order of keys being unimportant in
most cases, yet, can vary), we parse the output of our test programs
with python, and compare the decoded representations instead.

This has the downside of requiring PyMongo >= 1.9 for the test suite,
but I can live with that.