mongo_wire: insert & update commands
Started implementing the mongodb wire protocol: for now, we can
construct insert and update commands.

The API isn't all that friendly, but it's not meant to be: there will
be a utility layer on top of it, which will also connect to the
network, and accept responses too.

At least, that is the plan at this moment.