mongo-client: Add a request_id to the mongo_connection structure.
To make wrappers easier, and not require a manual request ID setting
all the time, store the last command's request ID within the
connection structure.

It is initialised to 0, and wrappers will be expected to check this
value, and put an incremented value into the packet they construct.

The connection's requestID field is updated by mongo_packet_send(),
just before returning successfully.