mongo-sync: Add skeleton code for SLAVE_OK support.
In order to properly support SLAVE_OK with the mongo_sync_ family of
functions, we need to store the setting on a per-connection basis.

To maintain separation between the base client and mongo-sync,
mongo_connection has been subclassed, and the mongo_sync_ functions
will use that type from now on.