mongo-private: New header for internal stuff.
There are a few things which need to be shared between the various
layers, but that should not be made public, since they don't make much
sense, or their entire purpose is bridging between two layers.

These will be collected in libmongo-private.h, and their
implementation will be in the layer where it makes most sense.

Right now, the _mongo_connection struct and mongo_connection_new()
were moved here. Both are used by the sync layer only, and should not
be exposed to applications.