mongo-sync: Add mongo_sync_reconnect().
Replace mongo_sync_connect_to_master() with a more flexible and
generic approach: mongo_sync_reconnect().

The new function will attempt to find an existing and available node
(either master or secondary, unless master is being enforced), making
sure that it pings, and if a master was requested, then that it is
indeed a master node.