mongo-sync: mongo_sync_reconnect() behaviour change.
Instead of returning a brand new connection object, overwrite the old
one (and return it then). This way, we can easily reconnect an
existing object without forcing the caller to explicitly prepare for
reconnections.

This is important, because the need to prepare would in the end
propagate back up to the calling application, and for those, we want
reconnects to be transparent.

This had a side effect of increasing the requests made towards the
server, thus, some tests needed to be updated.