tests: Blew up the old mongo-client test into smaller pieces.
The smaller pieces test specific funcionality, and do not try to
verify the result we get from the server: the purpose of the
mongo-client layer is to send and receive, interpretation is the job
of a higher layer.

Therefore, the new, split tests are considerably smaller than the old
suite, yet, they test more functionality due to their focus on doing
just that.