tests: Integrate the networked mongo_client tests into the unit tests.
Now that we have helpers that aid us in embedding networked tests
painlessly, do that, and embed them: no need to have them in separate
files anymore.

This improves maintainability, readability, and throughness, as one
file can test all aspects of a function, and the test is not split up
into different pieces.