mongo-sync: Move the connection check & reconnect code to its own function
Since most of the mongo_sync_cmd_ functions were doing connection
checking, and reconnecting, collected the whole lot into a new
function, _mongo_cmd_chk_conn().

This function either reconnects to a primary, if force_master is on,
or pings the current host and attempts to reconnect if it fails.

As a side effect, most errnos are properly propagated upwards now.