mongo-wire: Add mongo_wire_cmd_insert_n().
This new function is similar to the previous mongo_wire_cmd_insert(),
except that instead of taking a variable number of arguments, it needs
the number specified, and the documents are passed as an array.

This form is preferable when one does not know the number of documents
at compile time, but can count them at run-time, and still wants to
use bulk-insert.