sync: Add mongo_sync_connection->rs.seeds.
The rs.seeds list will contain the user-supplied list of replicaset
seeds, which will not change during the lifetime of the connection
object, nor will it be replaced upon reconnect, like rs.hosts is.

The purpose of rs.seeds is to provide a seed list, for cases where we
need to reconnect, but rs.hosts isn't populated yet.