sync: Wire in the replica-set seeding.
Upon connecting, add the host to the seed list and to the host list
aswell.

When adding a seed, prepend it to the host list aswell, so we don't
have to care about the seed list in reconnect.

When replacing a connection (as part of reconnect), populate the host
list with the contents of the seed list.