sync: Safe mode checking for insert & update commands.
When safe mode is enabled, check the result of getLastError() after
insert and update commands, and act accordingly.

As a consequence, fixed a few test cases that were misbehaving, and
did not pass safe mode checks.

Further tests that verify that safe mode works as expected will be
added later.